% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmaBase.R
\name{ArmaBase}
\alias{ArmaBase}
\title{Product of two matrices}
\usage{
ArmaBase(mat_1, mat_2)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
default matrix product operator from Armadillo C++ library.
}
