% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArmaColumnRow.R
\name{ArmaColumnRow}
\alias{ArmaColumnRow}
\title{Product of two matrices}
\usage{
ArmaColumnRow(mat_1, mat_2)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
for method in rows/columns from Armadillo C++ library.
}
