% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelColumn.R
\name{ParallelColumn}
\alias{ParallelColumn}
\title{Product of two matrices}
\usage{
ParallelColumn(mat_1, mat_2, t = 1)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
parallel for method in columns from parallel R library.
}
