% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelRow.R
\name{ParallelRow}
\alias{ParallelRow}
\title{Product of two matrices}
\usage{
ParallelRow(mat_1, mat_2, t = 1)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
parallel for method in rows from parallel R library.
}
