% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppOpenMP.R
\name{RcppOpenMP}
\alias{RcppOpenMP}
\title{Product of two matrices}
\usage{
RcppOpenMP(mat_1, mat_2, t)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
OpenMP backend from Rcpp R library interface.
}
