% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppParallel.R
\name{RcppParallel}
\alias{RcppParallel}
\title{Product of two matrices}
\usage{
RcppParallel(mat_1, mat_2, t = 1)
}
\arguments{
\item{mat_1}{First one matrix}

\item{mat_2}{Second one matrix}
}
\value{
A matrix with matrix product result
}
\description{
This function reads two matrices and operate the product between them, by using
RcppParallel backend and applying parallelFor for method in rows from RcppParallel R library.
}
